requirements.txt
# Core ETL dependencies
pandas==2.1.4
requests==2.31.0
pyyaml==6.0.1

# Testing
pytest==7.4.4
pytest-cov==4.1.0

# SQLite is part of the Python standard library — no package needed
# Faker for seeding mock data
Faker==22.0.0
